fixed bugs on linux
