+++ FreeTAKServer/components/extended/emergency/controllers/emergency_on_controller.py
@@ -18,11 +18,9 @@ from ..configuration.emergency_constants import (
     EMERGENCY_ON_BUSINESS_RULES_PATH,
     EMERGENCY_ALERT,
     BASE_OBJECT_NAME,
-    DEST_SCHEMA,
-    DEST_CLASS,
-    MAXIMUM_EMERGENCY_DISTANCE
 )
 
+from .emergency_general_controller import EmergencyGeneralController
 
 config = MainConfig.instance()
 
@@ -58,6 +56,8 @@ class EmergencyOnController(DefaultBusinessRuleController):
             # the DefaultBusinessRuleController evaluate_request
             internal_action_mapper=emergency_action_mapper,
         )
+        self.emergency_general_controller = EmergencyGeneralController(request, response, sync_action_mapper, configuration)
+        self.emergency_general_controller.initialize(request, response)
 
     def execute(self, method=None):
         getattr(self, method)(**self.request.get_values())
@@ -70,46 +70,6 @@ class EmergencyOnController(DefaultBusinessRuleController):
         with tracer.start_as_current_span("convert_dict_to_node") as span:
             span.add_event("adding remark to emergency on")
             self.request.get_value("model_object").detail.remarks.text = "CALL 911 NOW"
-
-    def retrieve_users(self) -> dict:
-        """get the available users"""
-        with open(config.UserPersistencePath, "rb") as f:
-            return pickle.load(f)
-
-    def add_user_to_marti(self, emergency: Event, user: Event):
-        """create a new marti dest for the given user to the provided emergency"""
-        self.request.set_value("object_class_name", DEST_CLASS)
-        configuration = self.request.get_value("config_loader").find_configuration(DEST_SCHEMA)
-
-        self.request.set_value("configuration", configuration)
-        new_dest = self.execute_sub_action("CreateNode").get_value("model_object")
-        new_dest.callsign = user.detail.contact.callsign
-        emergency.detail.marti.dest = new_dest
-
-    def get_model_object_from_user(self, user) -> Event:
-        if hasattr(user, "modelObject"):
-            return user.modelObject
-
-        elif hasattr(user, "m_presence"):
-            return user.m_presence.modelObject
-
-    def filter_by_distance(self, emergency: Event):
-        """filter who receives this emergency based on their distance from the emergency"""
-        self.users = self.retrieve_users()
-        for _, user_obj in self.users.items():
-            user_obj = self.get_model_object_from_user(user_obj)
-            user_point = user_obj.point
-
-            # check that the distance between the user and the emergency is less than 10km
-            # TODO: this hardcoded distance should be added to the business rules
-            if (
-                distance.geodesic(
-                    (user_point.lat, user_point.lon),
-                    (emergency.point.lat, emergency.point.lon),
-                ).km
-                < MAXIMUM_EMERGENCY_DISTANCE
-            ):
-                self.add_user_to_marti(emergency, user_obj)
 
     def parse_emergency_on(self, config_loader, tracer: Tracer, **kwargs):
         """this method creates the model object outline and proceeds to pass
@@ -139,8 +99,8 @@ class EmergencyOnController(DefaultBusinessRuleController):
 
             response = self.execute_sub_action("DictToNode")
 
-            # TODO: this should probably be moved out to a business rule call
-            self.filter_by_distance(response.get_value("model_object"))
+            self.emergency_general_controller.initialize(self.request, self.response)
+            self.emergency_general_controller.filter_by_distance(response.get_value("model_object"))
 
             for key, value in response.get_values().items():
                 self.response.set_value(key, value)
