+++ FreeTAKServer/components/extended/emergency/controllers/emergency_sender_controller.py
@@ -5,6 +5,11 @@ from digitalpy.core.zmanager.response import Response
 from digitalpy.core.zmanager.action_mapper import ActionMapper
 from digitalpy.core.digipy_configuration.configuration import Configuration
 
+from .emergency_general_controller import EmergencyGeneralController
+from ..configuration.emergency_constants import (
+    EMERGENCY_ALERT,
+    BASE_OBJECT_NAME
+)
 
 class EmergencySenderController(Controller):
     """this class is responsible for transmitting emergencies, however it is not used as all routing
@@ -17,9 +22,11 @@ class EmergencySenderController(Controller):
         response: Response,
         emergency_action_mapper: ActionMapper,
         configuration: Configuration,
+        sync_action_mapper: ActionMapper,
     ):
         super().__init__(request, response, emergency_action_mapper, configuration)
-
+        self.emergency_general_controller = EmergencyGeneralController(request, response, sync_action_mapper, configuration)
+    
     def execute(self, method=None):
         getattr(self, method)(**self.request.get_values())
         return self.response
@@ -44,17 +51,16 @@ class EmergencySenderController(Controller):
                 f"error broadcasting emergency {error}"
             )
 
-    def send_emergencies_to_client(self, **kwargs):
+    def send_emergencies_to_client(self, config_loader, user, **kwargs):
         """this method will broadcast all emergencies to a specific client"""
         emergencies = self.execute_sub_action("GetAllEmergencies").get_value(
             "emergencies"
         )
+        self.request.set_value("object_class_name", BASE_OBJECT_NAME)
+        self.emergency_general_controller.initialize(self.request, self.response)
         model_objects = list(emergencies)
-        messages = []
-        for emergency in list(emergencies):
-            # for each emergency we need to serialize it to the desired format
-            self.request.set_value("model_object", emergency)
-            sub_response = self.execute_sub_action("SerializeEmergency")
-            messages.append(sub_response.get_value("serialized_message"))
+        configuration = config_loader.find_configuration(EMERGENCY_ALERT)
+        self.request.set_value("configuration", configuration)
+        for emergency in model_objects:
+            self.emergency_general_controller.add_user_to_marti(emergency, user)
         self.response.set_value("model_object", model_objects)
-        self.response.set_value("serialized_message", messages)
